requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
arrow==1.2.3
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
carla==0.9.14
certifi==2023.7.22
charset-normalizer==3.2.0
clearml==1.13.1
click==8.1.7
cmake==3.27.5
contourpy==1.1.1
croniter==1.4.1
cycler==0.11.0
dateutils==0.6.12
deepdiff==6.5.0
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.12.4
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.2
furl==2.1.3
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.37
h11==0.14.0
huggingface-hub==0.17.3
idna==3.4
importlib-resources==6.1.0
inquirer==3.1.3
iopath==0.1.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lightning==2.0.9
lightning-cloud==0.5.38
lightning-utilities==0.9.0
lit==17.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.1.78
ordered-set==4.1.0
orderedmultidict==1.0.1
packaging==23.1
pandas==2.0.3
pathlib2==2.3.7.post1
Pillow==10.0.1
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
portalocker==2.8.2
psutil==5.9.5
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
PyJWT==2.4.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.9
pytz==2023.3.post1
PyYAML==6.0.1
readchar==4.0.5
referencing==0.30.2
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.10.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
termcolor==2.3.0
threadpoolctl==3.2.0
timm==0.9.7
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.2.0
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.10.1
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
wcwidth==0.2.6
websocket-client==1.6.3
websockets==11.0.3
yacs==0.1.8
yarl==1.9.2
zipp==3.17.0
